added controller, service and dto for tournament, changed tournament money related values from int64 to float64 and made the necessary db migration

+++ models/tournament.entity.go
@@ -23,8 +23,8 @@ type Tournament struct {
 	Description     string           `json:"description" gorm:"type:text"`
 	Status          TournamentStatus `json:"status" gorm:"type:varchar(20);default:'draft'"`
 	GameID          *uuid.UUID       `json:"game_id,omitempty" gorm:"type:uuid;index"`
-	EntryFee        int64            `json:"entry_fee" gorm:"default:0"`
-	PrizePool       int64            `json:"prize_pool" gorm:"default:0"`
+	EntryFee        float64          `json:"entry_fee" gorm:"default:0"`
+	PrizePool       float64          `json:"prize_pool" gorm:"default:0"`
 	MaxParticipants int              `json:"max_participants" gorm:"not null"`
 	StartsAt        time.Time        `json:"starts_at" gorm:"not null"`
 	EndsAt          time.Time        `json:"ends_at" gorm:"not null"`
@@ -41,9 +41,9 @@ type TournamentParticipant struct {
 	ID           uuid.UUID `json:"id" gorm:"type:uuid;primaryKey;default:gen_random_uuid()"`
 	TournamentID uuid.UUID `json:"tournament_id" gorm:"type:uuid;not null;uniqueIndex:idx_tournament_user"`
 	UserID       uuid.UUID `json:"user_id" gorm:"type:uuid;not null;uniqueIndex:idx_tournament_user"`
-	Score        int64     `json:"score" gorm:"default:0"`
+	Score        float64   `json:"score" gorm:"default:0"`
 	Rank         int       `json:"rank" gorm:"default:0"`
-	PrizeWon     int64     `json:"prize_won" gorm:"default:0"`
+	PrizeWon     float64   `json:"prize_won" gorm:"default:0"`
 	JoinedAt     time.Time `json:"joined_at" gorm:"autoCreateTime"`
 
 	//relationships
